fqdn: allow networks to isolate and managed name resolution
If HCP_FQDN_PATH is defined, a home-brewed FQDN publication and discovery
mechanism runs in all the defined container images. The mechanism relies
on a shared volume being mounted at that path in all the containers.

This is used to provide a closed-system where real-world FQDNs and
configurations can be used without conflicting with the real world. It
suppresses the use of docker hostnames for inter-container addressing,
and explicitly-defined FQDNs for the containers are used instead.

The configuration in docker-compose.yml and defaults.env is updated to
use this mode of operation.

Signed-off-by: Geoffrey Thorpe <geoffrey@twosigma.com>