enrollsvc: improved gencert and use of 'profiles'
This change incorporates updates in the safeboot submodule that allow
multiple, purpose-specific host certificates to be enrolled.

If the call to the enrollment API (/v1/add) provides a 'profile'
parameter that begins with "GENPROGS:", it will be used to set the bash
array variable of the same name for safeboot's 'attest-enroll' script,
thus controlling which (additional) assets get produced at enrollment
time.

Also, fix a possible leak in temporary file creation by adding an exit
trap to op_add.sh.

Signed-off-by: Geoffrey Thorpe <geoffrey@twosigma.com>