testcreds: default-on option to install "certissuer" trust root
The building of hcp_base includes an optional stage to embed a set of
host-provided trust roots. This extends it with a second option, to embed
the autogenerated "certissuer" testcred and have it be pre-trusted on all
derived images.

This also inverts the 3rd and 4th stages of "hcp_base", so that the
installing of packages occurs prior to the injection of trust roots. The
dependency to build testcreds can therefore use the former, and run
before (and be depended upon by) the latter.

Also, 'xxd' must be explicitly be installed, rather than relying on 'vim'
being installed to bring it in by dependency. ('vim' is an option, 'xxd'
is required by safeboot.)

Signed-off-by: Geoffrey Thorpe <geoffrey@twosigma.com>