build: remove excessive dependency
The point was to make sure that a tarball is rebuilt if whenever the list
of elements changes (even if the new or removed items weren't any newer
than the tarball). But this is handled by the dependency on a "deps" file
(which is itself rewritten during preprocessing if and only if it changes)
so we don't need a dependency on the _Makefile_ that writes out that deps
file!

Signed-off-by: Geoffrey Thorpe <geoffrey@twosigma.com>