build: robust dependency handling for images and Dockerfiles
This now uses the same approach as was already used for building code
tarballs. Namely, the 'make' preprocessing recalculates deps files but
only rewrites them into the FS if they've changed, then we include those
deps files, and one of the entries in that deps file declares a
dependency on ... the deps file. In this way, things get regenerated not
just when dependent inputs are newer, but when the _list_ of dependent
inputs changes.

Note that dependencies must be created at preprocessor time in order to
be included and considered, so we explicitly create $(HCP_OUT) during
preprocessing rather than having it be dependency-driven via MDIRS.

Signed-off-by: Geoffrey Thorpe <geoffrey@twosigma.com>