caboodle: significant UX improvements
Overhaul the shell interface for caboodle, and reduce maintenance overhead
for the list of mount points (that need to be manually created in caboodle
because it doesn't get those mounts).

Also, update caboodle_test.sh accordingly.

Also, acknowledge that the "networked" version of caboodle and its "soak"
handling needs to be reviewed and reworked.