usecase: renaming, consistency
Clarifies an approach that was already halfway in effect. The enrollsvc,
attestsvc, and attestclient definitions are named accordingly, whereas the
instantiations (services in docker-compose, volumes, usecase entities and
settings, etc) are intentionally cropped to distinguish them from the
implementation's ways and means. So for example we have emgmt and erepl
instead of enrollsvc_mgmt and enrollsvc_repl, and aclient instead of
attestclient.

Signed-off-by: Geoffrey Thorpe <geoffrey@twosigma.com>