enrollsvc: combine common.sh and common_defs.sh
These had been split out in order to take a subset of definitions and
functions, that deal with the enrolldb layout, and embed that code in the
enrolldb itself. I.e. so the attestsvc could interrogate the enrolldb
using the same definitions as the enrollsvc was using, and in theory this
would allow the layout to change. As with the versioning support, this
didn't really pan out and is now just complexity with no benefit, so it is
removed.

Signed-off-by: Geoffrey Thorpe <geoffrey@twosigma.com>