build: use attestation to trust testcred CA
This gets rid of the circular-dependency-workaround in hcp/base/ for
installing the testcred CA, and also means that testcred-generation has
access to all "tpmware" sw (including Heimdal), whether it's from OS
packaging or built-from-source.

Instead, the attestation client now supports a callback to automatically
install a 'certissuer.pem', if it shows up in the asset-bundle
post-attestation, as a trust root in the system.

* safeboot and install PATH/LD_LIBRARY_PATH are now added automatically
  when /hcp/common/hcp.sh is included, rather than by function call.
* Heimdal's hxtool is now used to produce the testcred CA
  ("enrollcertissuer"), which puts the key and cert in the same file.
  Adjust files/paths accordingly.
* Update caboodle's cred-generation to use hxtool also.

Signed-off-by: Geoffrey Thorpe <geoffrey@twosigma.com>