enrollsvc: improve environment-passing, remove "paramfile"
The JSON input ("profile") to the enrollsvc "add" API now allows an
arbitrary set of environment variables to be passed, though they all get
prefixed with "ENROLL_". As such, the contract on all genprogs is to be
wary of all variables of that form - without having to worry about
variables that don't have that prefix.

This generalization of the "profile" parameter makes the (still unused)
"paramfile" parameter increasingly pointless, so this removes it.

Signed-off-by: Geoffrey Thorpe <geoffrey@twosigma.com>