attestsvc: remove flask_wrapper.sh, use uwsgi priv-drop
Also, mount a static "ahcp.uwsgi.ini" (in /usecase) rather figuring out
the uwsgi configuration dynamically during start-up.

Also, improve healthcheck similar to what was done for enrollsvc. In the
case of attestsvc, the basic web API is provided by the flask app object
in safeboot's sbin/attest-server, so we need to declare a local wrapper of
that in order to inject the extra (healthcheck) handler.

Signed-off-by: Geoffrey Thorpe <geoffrey@twosigma.com>