source-building: define HCP_SOURCE_RESULT_$N
This simplifies the source-building Makefile itself, but also provides a
much easier handle to work with for code that uses this.

Also, mount source at /src-<name> rather than /<name>. Otherwise things
fail if you try to build and install a package to a path of the same name.
(Hint: safeboot)

Signed-off-by: Geoffrey Thorpe <geoffrey@twosigma.com>