ext-builder: used to be known as "source-builder"
This is for building and packaging code that's not an HCP app, but
something they depend on that aren't obtainable by system-native packaging.
Eg. tweaked versions of upstream codebases (eg. heimdal), other git
submodules (eg. safeboot), or even local code that we prefer to keep out
of the app code (eg. xtra).

Signed-off-by: Geoffrey Thorpe <geoffrey@twosigma.com>