orchestrator: support enroll_api.py HTTPS controls
- Setting HCP_CERTCHECKER to a path will cause "--cacert" to be used.
- Setting HCP_CERTCHECKER to "none" will cause "--noverify" to be used.
- Setting HCP_CLIENTCERT will cause "--clientcert" to be used.

Note, having trust roots and appropriately-sculpted client certs
provisioned and installed is what HCP does, but there's a catch. For a
host to get such things via attestation, an enrollment must have been
performed. The 'orchestrator' must be able to hit the enrollsvc::mgmt API
from an environment that isn't using a TPM and has not yet attested (even
if it's just in order to enroll a host+TPM that could do future
enrollments from a properly-attested environment).

Signed-off-by: Geoffrey Thorpe <geoffrey@twosigma.com>