docker: use --init to have process and signal mgmt in containers
The equivalent for docker-compose is to set the "init: true" option. This
causes containers to get a PID1 process that handles reparenting orphaned
child processes and signals from the host. The code was implementing its
own version of this so that can now be removed (or in the case of
launcher.sh, made unreachable but kept around for posterity).

Also, ahcp.uwsgi.ini needs the "die-on-term" setting, which emgmt already
had.

Signed-off-by: Geoffrey Thorpe <geoffrey@twosigma.com>