enrollsvc: fix for rootless docker and healthcheck
When running in exotic docker configurations (eg. rootless), the runtime
may not support containers listening on privileged ports, even on a
private network. A previous change allowed webservers to drop-privs after
binding sockets so that default ports (ie. 80 and 443) could be used, but
it's better if things work out of the box on rootless docker, for
dev/debug at least.

Also, fix healthcheck so that it works in both default and pruned
configurations.

Signed-off-by: Geoffrey Thorpe <geoffrey@twosigma.com>