policysvc: use a web-end point (not python callback) for policy
The idea of using a python file in /usecase as the "hook" for operators to
implement policy controls was already a stop-gap. Turns out, on rootless
docker that can run into trouble when the caller (in the container) is
running with dropped-privs. As such, we move to the proper model.

enrollsvc can have a policy variable set, which is a URL to a web API,
that causing it to consult that API when processing enrollment API
requests.

policysvc is a new service that serves as a stub. It implements the policy
control as a template/reference, but has no configurability to allow it to
impose any constraints. So it's no good for production. It will log some
output for all requests _except_ healthchecks (which are silenced).

Signed-off-by: Geoffrey Thorpe <geoffrey@twosigma.com>