caboodle: make sure CABOODLE_ALONE doesn't use the FQDN-BUS
Spotted this bug when running tests in parallel (which were using the same
project ID, an independent bug that was fixed in a previous bug). Easiest
fix is to just have CABOODLE_ALONE-themed containers to ignore any mount
for "FQDN_BUS" and just divert its own attention into a private directory.

Signed-off-by: Geoffrey Thorpe <geoffrey@twosigma.com>