HcpJsonPolicy: return more information from the filtering
Previously, the return value from the HcpJsonPolicy::run() call was
"accept" or "reject". Instead, we now return a dict that has that same
text outcome in it, but also provides the label of the final policy rule
that determined the outcome. If a legit operation is getting rejected, or
conversely an operation that should fail is being allowed, having a
labelled way to identify which of the policy entries made the final
determination may help to debug the policy configuration.

Signed-off-by: Geoffrey Thorpe <geoffrey@twosigma.com>