enrollsvc: upgrade gencert-hxtool to python
General cleanup and improvement of semantics occur along the way, and one
structural change that affects configuration: previously the
"gencert-hxtool" in a JSON profile would have a "list" field which was a
string containing a space-separated list of assets to produce. The
conversion to python3 highlighted that this isn't ideal. The natural way
to convey such a list is with ... a list. So the affected inputs in
usecase/ are updated, as is all the affected code. Existing installations
will need to update their JSON for enrollsvc, policysvc, and orchestrator.

Signed-off-by: Geoffrey Thorpe <geoffrey@twosigma.com>