policysvc: fix case where mounted files are only root-visible
The policysvc runs in uwsgi, which drops privs before handling traffic. In
some docker configurations (esp. "rootless"), files in host mounts aren't
world-readable. That's what happens with the "usecase/policy.json" file
that policysvc reads. Quick and nasty solution: have the startup script
copy it to a temp file, chmod the temp file to be world-readable, and
repoint the HCP_POLICYSVC_JSON env-var to point to that temp file.

Signed-off-by: Geoffrey Thorpe <geoffrey@twosigma.com>