gencert-hxtool: support propval arrays and dicts
Currently, the gencert-hxtool structure (dict) in a profile consists of
key-value pairs that translate to command-line arguments to hxtool. This
poses the challenge of how to send the same argument multiple times?
Answer: if the value is actually an array of values, to expand it to
multiple occurrences of the same key with the different values. (As a
bonus, if the value is a dict, assume that it should be converted to a
JSON string value instead.)
command-line arg.

Signed-off-by: Geoffrey Thorpe <geoffrey@twosigma.com>