genreenroll: add a "hint-reenroll-<expiry>" to enrollments
This generates a JSON file about when the enrollment occurred and when it
should be reenrolled. The formatting of the filename is
"hint-reenroll-YYYYMMDDhhmmss", where the datetime information refers to
when the enrollment should be scheduled for reenrollment. This is so that
an ordered (sorted) listing of such files (using a glob) can identify the
the enrollments need enrolling without reading/parsing their content. The
JSON content itself uses python's 'datetime.isoformat()' method to format
the last enrollment time, the next enrollment time, and as a third field
it also stores the number of seconds interval between the two.

Signed-off-by: Geoffrey Thorpe <geoffrey@twosigma.com>