tests: wait for emgmt to be up when it is a dependency
In extreme environments (we saw this in a CI pipeline), it may take so
long for emgmt to be ready that the orchestrator gives up. (In the default
config, emgmt attests to get its server certificate, so this creates a
dependency on the full initialization flow. And 'orchestrator' is a client
tool, not a service, so it can't spin/wait forever.

Also, add some annotation to the test output.

Signed-off-by: Geoffrey Thorpe <geoffrey@twosigma.com>