enrollsvc: use http status codes properly and improve logging
Prior to this change, the success/failure of APIs was only partially
correlated to the http status code returned with the response, requiring
the caller to inspect the response body to make conclusions. This corrects
that.

This change also directs much of the enrollsvc code to use the
hcp_tracefile abstraction for logging output.

- db_*** scripts, used to implement the individual operations for the
  API, now convey an http status code as their exit code, rather than the
  conventional posix exit codes (where 0 is success).
- more (and better) logging is put in place.

Signed-off-by: Geoffrey Thorpe <geoffrey@twosigma.com>