HcpRecursiveUnion: dedup for union of JSON lists
Eg. suppose you are taking the union of two JSON docs and they each
have their own version of a "task list" field. If there are some tasks
that are listed in both documents, you probably don't expect the union
to have two copies of them, so listDedup defaults to True. If either of
the documents already has duplicates in its respective field then this
de-duplicating behavior will that too, it isn't specific to duplication
created across the two docs.

Signed-off-by: Geoffrey Thorpe <geoffrey@twosigma.com>