enrollsvc: structure change for enrollsvc.json
The existing "preclient" and "postclient" elements are specific to the
db_add (ie. enrollment and reenrollment) functions, so we add a "db_add"
field and move preclient/postclient under that. (We will soon wish to add
config to other aspects of enrollsvc, so this creates some elbow room.)

Also, hide a vim modeline in the JSON (as the value of a "_" field).

Signed-off-by: Geoffrey Thorpe <geoffrey@twosigma.com>