HcpJsonPath: improvements and new features
Path "nodes" are simply JSON object (and python dict) keys, which are
arbitrary strings, so the only real character limitation we _need_ to
enforce is to prohibit '.' because we use that as a path-separator and
really don't want to get into the business of escaping. However we've
gotten by so far with very restrictive path nodes ([A-Za-z0-9_]) so this
just allows some extra characters that we know we want ([A-Za-z0-9_<>-]).

Supplement the existing extract_path() API with;
 - overwrite_path(), which puts an arbitrary value at the given path,
   insisting that the parent node therefore be a dict. This is destructive
   if it needs to be, and adds missing paths/fields if required.
 - delete_path(), removes whatever is at the given path if it exists.
   Note, there is no automatic cleanup of any fields left empty by this
   operation. Ie. the creation of missing fields by overwrite_path() is
   not be undone by delete_path().

Signed-off-by: Geoffrey Thorpe <geoffrey@twosigma.com>