HcpRecursiveUnion: fix list-deduplication
When doing a deep-merge/union of two structures, two lists (at the same
path) will be joined and then de-duplicated, unless the 'listDedup=True'
parameter is overriden. We had been using a technique for doing this that
is widely-repeated on the internet, namely converting the combined list to
a 'dict' (with empty values) and then back to a list, thus relying on dict
keys being unique. This technique has a severe limitation though, namely
that all of the items in the lists must be 'hashable'. E.g. if the list
has dict elements, this fails. We now use a more reliable loop construct.

Signed-off-by: Geoffrey Thorpe <geoffrey@twosigma.com>