enrollsvc: remove self-enrollment and integrated-TPM
This workflow proved "too cunning to be understood". Instead we add
another "testcred" to generate and use in the dev/debug workflow, knowing
that a real deployment would be plugging in its own creds. Self-enrollment
was a smarty-pants proof-of-concept, and a good built-in test-case, but
clarity is the best thing now.

Signed-off-by: Geoffrey Thorpe <geoffrey@twosigma.com>