attest_callback_common: add a python version with more tricks
The bash callback remains a likely candidate for use (or reference) for
constrained environments, where a shell is easier to justify than a full
python installation. But for the "refresh" workflow, where the attestation
client runs periodically in the host's OS, python is less onerous. It's
also easier for adding some smarts that we want in such a use-case, like
the ability to run callbacks when relevant assets are updated. (Eg. to
restart services only when the credentials refresh.)

Signed-off-by: Geoffrey Thorpe <geoffrey@twosigma.com>