tests: rejig the healthchecks and better failure handling
The idea of having the healthcheckers be general tools in /hcp/tools was
short-sighted, because unless they are always run inside their respective
containers they are easily broken by the slightest change. So we return
them to their respective codebases. Tests and scripts that need to
healthcheck (and possibly poll) those containers should instead do so by
running "docker-compose exec <container> <healthcheck>" directly, which is
what this change does for the test code.

Also, the idea of executing tests/wrapper.sh and having that _source_ the
test case proved a little unstable. (Implementing a trap in a test would
silently override the wrapper's trap, leading to bad leaks.) This has now
been fixed.

Finally, run_orchestrator was not propagating in-container errors, but
"set -e" fixes that.

Signed-off-by: Geoffrey Thorpe <geoffrey@twosigma.com>