kdcsvc: add a healthcheck
This healthcheck for kdcsvc bases itself on hitting the /healthcheck web
API with curl, like the enrollsvc::mgmt and attestsvc::hcp services. In an
ideal world, we'd also check that the KDC itself is functioning (eg. by
running a "kinit -C FILE:/etc..... <princ> klist" command), but that's for
another time. (Our current use of healthchecks is primarily to wait for
them to be "up" - and for kdcsvc the web API is started after the KDC.)

Signed-off-by: Geoffrey Thorpe <geoffrey@twosigma.com>