kdcsvc: use very short key-rotation cycles
Rather than measuring the KDC's key-rotation schedule in days, we measure
it in seconds. This is to put more heat on the whole credential management
cycle, from reenrollment, to periodic re-attestation, to HUP'ing cred-
dependent services (I'm looking at you, openssh).

Eg. this quickly breaks code that runs "kinit" once and relies on the TGT
in the cred-cache. This is why we tend to run "kinit [...] <cmd>", with
where the kinit stays resident for that command and automatically
refreshes the credential cache.

Signed-off-by: Geoffrey Thorpe <geoffrey@twosigma.com>