caboodle: revamp 'caboodle_networked' and give it a TPM
The caboodle_networked container is now a steady-state, "do_nothing"
machine whose job is to stay alive and idle, like a regular server, so
that arbitrary actions can be performed on it (using 'docker-compose exec')
that retain state from one action to the next.

This also adds a healthcheck, that verifies if it can obtain a TGT. This
implicitly makes the health status dependent on successful enrollment and
attestation (without which the 'pkinit-client' certificate used to obtain
a TGT won't be present).

This necessarily updates the enrollsvc policy configuration, in which we
allow 'caboodle_networked' and 'sherver' to get pkinit-client certificates
for arbitrary principals. This is not necessarily a good or bad idea, and
it's not production - this just gives us a different config from the other
hosts, which is a more interesting test-case (and more flexibile).

Signed-off-by: Geoffrey Thorpe <geoffrey@twosigma.com>