build: support image dependencies outside "apps"
This requires that other build definitions use the same _DNAME and _TFILE
naming consistently, so that dependencies can be generated by symbolic
means. The app_image() function is enhanced to support this when the
'ancestor' parameter is of the form 'SYM:FOO'; the image will be derived
from $(HCP_FOO_DNAME) and the dependency will be on $(HCP_FOO_TFILE).

Signed-off-by: Geoffrey Thorpe <geoffrey@twosigma.com>