launcher: support "exec-<foo>" and "lights-out" for 'exec'
In the launcher steps, we have "setup-*", "start-*", and "custom". We now
also have "exec-*", whose job it is to literally 'exec' to the requested
service, replacing the launcher process there and then. This is needed in
cases where the launched service needs to run as PID 1. (In which case,
launcher has to be launched as PID 1 too.)

If a 'lights-out' setting is provided, the launcher will 'exec' to it
rather than exiting.

Signed-off-by: Geoffrey Thorpe <geoffrey@twosigma.com>