sshsvc: use ktutil for handling updated keytabs
Prior to this change, the attestation-received keytab was replacing the
existing one (if they were different), but that bypasses any discipline
that might be required for concurrent usage/updates on the keytab file.
This change redirects incoming keytabs to a modified path (suffixed with
".updated"). The attestation callback is the sole producer to /consumer of
that file so there is no such concurrency problem. When the file is
updated, we use "ktutil copy" to migrate keys to the intended keytab, and
we also run "ktutil purge" to clear out anything old.

Signed-off-by: Geoffrey Thorpe <geoffrey@twosigma.com>