docker-compose: give useful hostnames to containers
For some time, we had been validating that the 'fqdn-bus' technique was
working (in terms of defining the FQDNs that hosts are expected to use to
reach each other) by the fact that our containers were coming up with
meaningless hex names and hostnames. We're further along now, and we have
bootstrapping scenarios of authentication infrastructure where interactive
logins are involved, so friendly hostnames are worth it.

Signed-off-by: Geoffrey Thorpe <geoffrey@twosigma.com>