usecase: remove port overlaps (esp. KDC), and tidier logs
In order to support the 'monolith' use-case, where all tools and services
run in the same container, we need to eliminate any cases where those
tools and services listen on the same ports in different containers, as
that will fail when they run co-tenant. The only difficult case is the
TGS service for the primary and secondary KDCs, which both listen on port
88 by default. As the primary TGS is only supposed to be accessed _by_
the secondaries (and then, only for bootstrapping iprop - after which the
secondary can do everything the primary can), we move that one to port 99.

Note, the most noisy/repetitive services for generating debug logs (client,
fqdn_updater, attester) have explicit "unset HCP_NOTRACEFILE" settings
added, so that if HCP_NOTRACEFILE is set more generally (so that debugging
converges on stderr), we still remove the noisy sources. (The reason is
that, otherwise, those sources occur so often that the tracing that is of
genuine interest will quickly scroll out of view.)

Signed-off-by: Geoffrey Thorpe <geoffrey@twosigma.com>