attester: complement 'period' with a 'retry' setting
It makes sense for the 'period' attribute to be long, as a successful
attestation serves to get cred/config updates, which may only be necessary
on periods measured in hours or days. But when bringing up services out of
order, if a first attestation fails (because attestsvc lost the start-up
race, or because TPMs aren't done being created and/or enrolled, ...) we
don't want to wait long before a retry.

Signed-off-by: Geoffrey Thorpe <geoffrey@twosigma.com>