restarter: new service-wrapper to support restart
If a service launched by the launcher exits, the launcher assumes it to
spell "the end" and it tears down everything else too. If (sub-)services
exit "easily" for reasons that should be tolerated (eg. outages of
required services), this wrapper can keep going and restart the service
after a configurable delay.

Signed-off-by: Geoffrey Thorpe <geoffrey@twosigma.com>