common: relocate HCP_CONFIG_FILE to the rootfs
Ideally the config file should be read-only, root-owned, and
world-readable, but "mount" options and exotic docker configurations can
cause any of those conditions to fail. So we twiddle the hcp_config
code (for bash and python) to determine, when loading, if the
HCP_CONFIG_FILE path is a child of /tmp/workloads - if it isn't, we
copy the file there, give it the desired attributes, and repoint the env
var to it. (This also integrates the pre-existing trick we had, for
the case where non-root processes don't inherit HCP_CONFIG_* environment
variables and can't/shouldn't trust anything in their inputs to figure
out what path to read config from -> we store the settings in the home
directory when creating the account.)

Signed-off-by: Geoffrey Thorpe <geoffrey@twosigma.com>