testcreds: use a new HCP_NO_CONFIG option
In the quirky case of testcreds, we are running specific short-lived
containers to generate some creds directly form 'hxtool' and friends. The
commands source /hcp/common/hcp.sh only to ensure that the PATH and so on
are set up properly. In this case, there is no HCP workload and no JSON
configuration.

Signed-off-by: Geoffrey Thorpe <geoffrey@twosigma.com>