config: allow config relocation to /etc
To avoid issues with exotic docker configurations, the workload JSON input
is relocated to /tmp and chmod'd for world-readability, if it hasn't
already happened. But there are situations (in this commit, webapi.sh)
where the workload JSON will be further modified and stored elsewhere,
meaning that the next time HCP notices it's not reading from /tmp it'll
relocate it ... Break the cycle by tolerating the case where the config
file is in /etc.

Signed-off-by: Geoffrey Thorpe <geoffrey@twosigma.com>