usecase: make exec/shell wrappers use launcher
Up till now, running 'exec' or 'shell' gave a raw shell with no control
over environment or scripting activity. Via docker-compose, the
environment for the spawned process would be whatever environment was set
when the container was created. Via monolith, the environment would be
that of the caller. Neither is ideal.

Now that workloads can supply global-scoped 'env' sections, we make the
wrappers for exec/exec_t/shell deliberately run bash under "launcher" so
that the environment gets set.

Signed-off-by: Geoffrey Thorpe <geoffrey@twosigma.com>