preproc: fix dependency bug in image building
We accumulate the list of locally-built package files in 'pkgs_local_src'
in order to hardlink them prior to "docker build" (and remove them after).
But the image build rule shouldn't be dependent on those files, but should
instead depend on their corresponding touchfiles (which aren't 1:1 with
package files - so duplication-detection is required).

Signed-off-by: Geoffrey Thorpe <geoffrey@twosigma.com>