builder: split builder into 2 layers to reduce churn
pp_add_layer() intentionally schedules the installation of locally-built
packages before external/upstream ones. This helps avoid an upstream
package automatically pulling (by dependency) another upstream package in
the case where we have a locally-built version. The problem is that any
local change may trigger expensive and needless download and reinstall
overheads. In the case of 'builder', we know that the upstream packages
listed have no dependency on the locally-built packages, so to better
benefit from docker (and buildkit) caching, we'd like the upstream
packages to install earlier.

Solution: split the 'builder' layer into 2 layers, doing the upstream
package installations in the first.

Signed-off-by: Geoffrey Thorpe <geoffrey@twosigma.com>