hcp: rename 'caboodle' code to 'monolith', leave image alone
The code in src/hcp/caboodle is to support the running of a single
container with multiple workloads running co-tenant, which is the
'monolith' model and the one used by the 'monolith' docker-compose target,
so the code (and resulting package) are renamed to 'monolith'.

The 'caboole' _image_ OTOH is the single application image we now build,
so named because it contains all the HCP code and can be used when running
any workload (including 'monolith', but everything else too). So we leave
the 'caboodle' name to represent this (and only this).

Signed-off-by: Geoffrey Thorpe <geoffrey@twosigma.com>