cache: implement caching of *.deb and *.tar.gz builds
If the desired packages are in cache/{debbuilder,builder}/, they are used,
irrespective of whether or not we have the source code and means to build
them and irrespective of how up-to-date (or otherwise) they are. If the
packages are not there, then we revert back to HCP building those packages
from source, and dependencies (as usual) automatically govern if and when
they need rebuilding.

"make cache_update" reverses the dependency/priority order, such that HCP
builds of the packages are required, and (once built) will replace the
packages in the cache if they are not binary equal.

Signed-off-by: Geoffrey Thorpe <geoffrey@twosigma.com>