hcp: by default, HCP source is now mounted, not installed
This removes the use of an 'hcp' package (via "builder") and instead all
HCP source is mounted into containers (and from there, into VMs).

This also necessarily fixes up a bunch of the UML vs QEMU dysfunction, and
a few bugs found along the way. As a result, there is now a "uml_runner"
equivalent to "qemu_runner" (though it still has some issues).

Signed-off-by: Geoffrey Thorpe <geoffrey@twosigma.com>