kdcsvc: add a "realm_healthcheck" interface
A "realm_healthcheck" subcommand is added to the do_kadmin.py utility, and
a distinct "realm_healthcheck.py" script is added to simplify calling the
do_kadmin.py subcommand (eg. from a docker-compose healthcheck or a test
script).

This will attempt to run a "kadmin -l ext_keytab" operation for the
"host/healthcheck.{domain}" principal. The specific purpose is to see
whether the realm data is sufficiently initialized on the local KDC, most
notably when the KDC is a secondary and replicating from a primary.

NB: this healthcheck does _not_ guarantee that the webapi or kdc_tgs
services are up and accepting incoming connections.

- usecase: 'kdc_secondary' allows realm_healthcheck.
- test_{sequential,parallel}: block on kdc_secondary's realm_healthcheck.
- docker-compose: add realm_healthcheck to both KDCs.
- remove the horrible workaround from genkrb5keytab.

Signed-off-by: Geoffrey Thorpe <geoffrey@twosigma.com>