debbuilder: support extra args for docker-run, and fix trap
When adding a codebase as a debbuilder target, <name>_PKG_CMD_DRUN_FLAGS
can be used to pass arbitrary cmd-line options to "docker run".

The container command run when bootstrapping and packaging a debbuilder
target sets a trap in order to chown any files the container might have
modified that reside in bind mounts from the host. But using ".." turned
out to be a bad path specified, because we aren't sure what the current
directory will be at the moment of exit. Change this to an explicit path.

Signed-off-by: Geoffrey Thorpe <geoffrey@twosigma.com>