uml/qemu: use VDE's DNS proxying
slirpvde hosts a DNS server on its own IP address, but you need to tell
its DHCP server to announce that. Within the VM side, we need to enable
systemd-resolved to pick up on the DNS information from DHCP and publish
a corresponding resolv.conf. Finally, we need to symlink /etc/resolv.conf
to that publication, but we can't do so from a Dockerfile (because docker),
so we force a symlink into the tarball that is extracted from the docker
image.

Signed-off-by: Geoffrey Thorpe <geoffrey@twosigma.com>