qemu: use "shutdown" rather than "/myshutdown"
The /myshutdown util provides an immediate (and brutal) termination of the
kernel. This is useful when running the HCP launcher as "init" in a VM
that uses a throwaway disk layer, because it gives the same behavior as a
container running the same workload - the launcher is PID 1, and when it
exits all running tasks terminate unceremoniously and all volatile state
vanishes.

But when mounting persistent filesystems and starting up backgrounded
services outside of HCP, it's better to do a controlled shutdown.

Signed-off-by: Geoffrey Thorpe <geoffrey@twosigma.com>