usecase: add 'target' and 'bigbrother' nodes
'bigbrother' acts as a "docker exec"-like jump host. It doesn't run any
services, and gets issued with a pkinit cert+key for the root principal.
It's an easy way to get a root shell on any host (that runs sshd). This is
useful for VMs (where docker-exec isn't available), and even when
docker-exec can be used, this is often preferable anyway, as it sets basic
env-vars (HCP_CONFIG_FILE, KRB5_CONFIG, etc) and gives more predictable
terminal behavior. It's also preferable to use bigbrother as an ssh client
than to ssh from the host, as the later requires coordinating identity
(and creds) inside and outside the usecase.

'target' is a pure ssh target, like sherver.

Signed-off-by: Geoffrey Thorpe <geoffrey@twosigma.com>