usecase: reduce expiry: certs to 1h, clockskew to 1s
pkinit-client certs (and TGTs) are already much more aggressive (at 10m
and 5m, respectively), but for other certs, a 1h expiry is still more
likely to show any life-cycle/refresh bugs in development than, say, 1y!

Also, heimdal's default clockskew-tolerance is 5m, which is enough to bend
the very life-cycle rules we're trying to validate with pkinit-cert and
TGT refresh, so we wind it way down to 1s. (Containers and VMs running in
the same host should be within microscopic tolerances.)

Signed-off-by: Geoffrey Thorpe <geoffrey@twosigma.com>