usecase: bigbrother can avoid to make its config be global
bigbrother is used as a root-credentialed ssh client to the SSO
environment, which works for VMs as well as containers (docker-exec only
gets us into the container), and ssh often provides saner terminal
semantics. There's no point having bigbrother in the 'monolith' model, so
there's no need for it to be a good co-tenant, and having its config be
global means there's no need to set KRB5_CONFIG/KTNAME, paths are easier,
etc.

Also, start an sshd in bigbrother - why not? (It helps to have another ssh
target when looking at delegation.)

Signed-off-by: Geoffrey Thorpe <geoffrey@twosigma.com>