attest_callback: implement a 'hostcerts' callback
This callback implementation allows rules to be implemented in a JSON
list, using regex matching on the asset name, that cause the asset to be
copied to a given destination with a given owner and mode. Critically, it
also handles the situation where there are rules specifying owners
(accounts) that don't (yet) exist.

Signed-off-by: Geoffrey Thorpe <geoffrey@twosigma.com>